// backend/src/core/payments/payment-reference-type.enum.ts
export enum PaymentReferenceType {
  SHOP_ORDER = 'shop_order',

  // C'EXPRESS
  EXPRESS_DELIVERY = 'express_delivery',
  EXPRESS_IMPORT_EXPORT = 'express_import_export',

  // prêts pour la suite
  EVENT_BOOKING = 'event_booking',
  CLEAN_BOOKING = 'clean_booking',
  TODO_TASK = 'todo_task',
  GRILLFOOD_ORDER = 'grillfood_order',
}